Release 11.15.02
git-svn-id: http://svn.code.sf.net/p/netpbm/code/advanced@5252 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/jbig/libjbig/jbig.c
@@ -928,7 +928,7 @@ void
 jbg_enc_options(struct jbg_enc_state *s,
                 int const order,
                 int const options,
-                unsigned long int const l0,
+                int const l0,
                 int const mx,
                 int const my) {
